Dev Updates (#3)
* fix: folder naming & scripts & removed unused items

* fix(db): for entities

* chore(formatDate): changed lang type

* feat(user/settings): notify after save

* chore: eslint setup

* fix(client): eslint warns & formatted code

* feat(client): boring-avatars

* feat(client): upgraded packages & removed temporally next-pwa

* refactor(client): for `Input`

* refactor(client): for `Toggle`

* efactor(client): for `Modal`

* fix(client/Header): for avatar

* refactor(client): removed unwanted

* fix(client): minor changes for styles

* style(client): improvements with look

* style(client): update for cards look

* style(client): overall update with colors & vars

* style(client): changed colors

* style: color palette change

* style(client): color palette changes

* fix(client): with styles

* fix(client/set): minor change for `h1`

* feat(client): refactor & updated styles & component Button

* fix: minor improvements & fixes for lint

* fix(client): overall improvements with look and feel

* fix(client): with AnimatePresence & styles

* fix(Modal): for `AnimatePresence` prop

* style(globals): update disabled selector for button

* Update For Flashcards (#4)

* feat(client): base for change flashcards

* feat(flashcards): updated animations

* feat(flashcards): added result info after completion

* fix: minor changes with styles and look

* fix(input): minor change

* style(client/set): update for whole page

* feat(client): updates with ui for mobile first & refactor

* feat(client): updates with ui for mobile first

* chore: updated readme

* chore: updates with SEO items

* fix(client/BottomNavigation): for nav item

* feat(client): updates with ui for mobile first

* refactor(client/AuthCheck): minor change

* fix(client/_app): changed title

* feat(client): updated home page

* fix(client): for exam page for styles

* feat(client): improvements with set creation

* fix(client/SetEditing): for cards

* feat(client): update for folder & set creation

* chore(client): corrected naming of folders

* feat(client): updates for error pages

* feat(client): update with access for set page

* fix(client): for sets & folders for getting their items

* style(client): fix for toggle

* feat(server): logic for logout

* feat(client): logic for logout

* feat(server): updates for authorization

* feat(client): updates for authorization

* fix(client/SetEditing): minor changes

* fix(server): for authorization

* feat(client): updates for authorization

* fix(server): updates for folder & set endpoints

* fix(client/FolderEditing): for after save behaviour

* feat(client): updates for folder & set apis; updated home page

* packages upgrade (#5)

* chore(server): upgraded packages to latest version

* refactor(server): typeorm migration updates

* fix(server): updates for services; wrapped to `try..catch` & added loggers

* chore: removed unwanted package & items

* fix(client): for folder save & delete

* fix(server): order sort for get sets & folders

* fix(client): in apis for errors handle

* fix(client): delete handle for set

* fix(client/set): minor changes with content

* fix(client): in apis for get folders & sets

* fix(client): in home page for content

* fix(server): in services for get folders & sets

+++ server/src/controllers/userController.ts
@@ -1,5 +1,6 @@
 import { Request, Response, NextFunction } from 'express';
 import { userService } from '@/services/userService';
+import { ReqWithSessionValues } from '@/interfaces';
 
 class UserController {
   public getUsers = async (_: Request, res: Response, next: NextFunction) => {
@@ -21,21 +22,27 @@ class UserController {
     }
   };
 
-  public updateUser = async (req: Request, res: Response, next: NextFunction) => {
+  public updateUser = async (req: ReqWithSessionValues, res: Response, next: NextFunction) => {
     try {
       const payload = req.body;
-      await userService.update(payload);
-      res.status(200).json({ message: 'updated' });
+      if (payload.id !== req.userId) res.status(401).json({ status: 403, message: 'Forbidden' });
+      else {
+        await userService.update(payload);
+        res.status(200).json({ message: 'updated' });
+      }
     } catch (error) {
       next(error);
     }
   };
 
-  public deleteUser = async (req: Request, res: Response, next: NextFunction) => {
+  public deleteUser = async (req: ReqWithSessionValues, res: Response, next: NextFunction) => {
     try {
       const payload = req.params.id;
-      await userService.delete(payload);
-      res.status(200).json({ message: 'deleted' });
+      if (payload !== req.userId) res.status(401).json({ status: 403, message: 'Forbidden' });
+      else {
+        await userService.delete(payload);
+        res.status(200).json({ message: 'deleted' });
+      }
     } catch (error) {
       next(error);
     }
